requirements.txt
pyyaml==6.0
pydantic>=1.10.7,<2.0.0
RPi.GPIO==0.7.1; platform_machine == 'armv7l' or platform_machine == 'aarch64'